Migrated to Qt5.
Compiled with Qt 5.0.2 MinGW 4.7 32bit.

+++ Common.h
@@ -50,7 +50,7 @@ typedef const char * LPCTSTR;
 #endif
 
 //typedef qint32 LONG;
-typedef quint32 DWORD;
+typedef long unsigned int DWORD;
 typedef quint16 WORD;
 typedef quint8 BYTE;
